refactor createPersonalAccountWithEmailViewModel to multiple viewModels
