+++ app/src/main/kotlin/com/waz/zclient/feature/auth/registration/personal/email/CreatePersonalAccountEmailViewModel.kt
@@ -1,0 +1,81 @@
+package com.waz.zclient.feature.auth.registration.personal.email
+
+import androidx.annotation.StringRes
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.waz.zclient.R
+import com.waz.zclient.core.exception.Failure
+import com.waz.zclient.core.exception.NetworkConnection
+import com.waz.zclient.shared.activation.usecase.EmailBlacklisted
+import com.waz.zclient.shared.activation.usecase.EmailInUse
+import com.waz.zclient.shared.activation.usecase.SendEmailActivationCodeParams
+import com.waz.zclient.shared.activation.usecase.SendEmailActivationCodeUseCase
+import com.waz.zclient.shared.user.email.ValidateEmailError
+import com.waz.zclient.shared.user.email.ValidateEmailParams
+import com.waz.zclient.shared.user.email.ValidateEmailUseCase
+import kotlinx.coroutines.Dispatchers
+
+class CreatePersonalAccountEmailViewModel(
+    private val validateEmailUseCase: ValidateEmailUseCase,
+    private val sendEmailActivationCodeUseCase: SendEmailActivationCodeUseCase
+) : ViewModel() {
+
+    private val _isValidEmailLiveData = MutableLiveData<Boolean>()
+    private val _sendActivationCodeSuccessLiveData = MutableLiveData<Unit>()
+    private val _sendActivationCodeErrorLiveData = MutableLiveData<ErrorMessage>()
+    private val _networkConnectionErrorLiveData = MutableLiveData<Unit>()
+
+    val isValidEmailLiveData: LiveData<Boolean> = _isValidEmailLiveData
+    val sendActivationCodeSuccessLiveData: LiveData<Unit> = _sendActivationCodeSuccessLiveData
+    val sendActivationCodeErrorLiveData: LiveData<ErrorMessage> = _sendActivationCodeErrorLiveData
+    val networkConnectionErrorLiveData: LiveData<Unit> = _networkConnectionErrorLiveData
+
+    fun validateEmail(email: String) {
+        validateEmailUseCase(viewModelScope, ValidateEmailParams(email), Dispatchers.Default) {
+            it.fold(::validateEmailFailure) { updateEmailValidationStatus(true) }
+        }
+    }
+
+    private fun validateEmailFailure(failure: Failure) {
+        if (failure is ValidateEmailError) {
+            updateEmailValidationStatus(false)
+        }
+    }
+
+    private fun updateEmailValidationStatus(status: Boolean) {
+        _isValidEmailLiveData.value = status
+    }
+
+    fun sendActivationCode(email: String) {
+        sendEmailActivationCodeUseCase(viewModelScope, SendEmailActivationCodeParams(email)) {
+            it.fold(::sendActivationCodeFailure) { sendActivationCodeSuccess() }
+        }
+    }
+
+    private fun sendActivationCodeSuccess() {
+        _sendActivationCodeSuccessLiveData.value = Unit
+    }
+
+    private fun sendActivationCodeFailure(failure: Failure) {
+        if (!isNetworkConnectionFailure(failure)) {
+            when (failure) {
+                is EmailBlacklisted -> _sendActivationCodeErrorLiveData.value =
+                    ErrorMessage(R.string.create_personal_account_with_email_email_blacklisted_error)
+                is EmailInUse -> _sendActivationCodeErrorLiveData.value =
+                    ErrorMessage(R.string.create_personal_account_with_email_email_in_use_error)
+            }
+        }
+    }
+
+    private fun isNetworkConnectionFailure(failure: Failure): Boolean {
+        return if (failure is NetworkConnection) {
+            _networkConnectionErrorLiveData.value = Unit
+            true
+        } else false
+    }
+}
+
+
+data class ErrorMessage(@StringRes val message: Int)
