+++ app/src/main/kotlin/com/waz/zclient/feature/auth/registration/personal/pincode/CreatePersonalAccountPinCodeFragment.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.feature.auth.registration.personal.email
+package com.waz.zclient.feature.auth.registration.personal.pincode
 
 import android.content.DialogInterface
 import android.app.AlertDialog
@@ -13,13 +13,15 @@ import com.waz.zclient.core.extension.sharedViewModel
 import com.waz.zclient.core.extension.showKeyboard
 import com.waz.zclient.core.extension.viewModel
 import com.waz.zclient.feature.auth.registration.di.REGISTRATION_SCOPE_ID
+import com.waz.zclient.feature.auth.registration.personal.email.EmailCredentialsViewModel
+import com.waz.zclient.feature.auth.registration.personal.name.CreatePersonalAccountNameFragment
 import kotlinx.android.synthetic.main.fragment_create_personal_account_pin_code.*
 
 class CreatePersonalAccountPinCodeFragment : Fragment(
     R.layout.fragment_create_personal_account_pin_code
 ) {
 
-    private val createPersonalAccountWithEmailViewModel: CreatePersonalAccountWithEmailViewModel
+    private val createPersonalAccountPinCodeViewModel: CreatePersonalAccountPinCodeViewModel
         by viewModel(REGISTRATION_SCOPE_ID)
 
     private val emailCredentialsViewModel: EmailCredentialsViewModel
@@ -47,21 +49,21 @@ class CreatePersonalAccountPinCodeFragment : Fragment(
 
     private fun initResendCodeListener() {
         createPersonalAccountPinCodeResendCodeTextView.setOnClickListener {
-            createPersonalAccountWithEmailViewModel.sendActivationCode(email)
+            createPersonalAccountPinCodeViewModel.sendActivationCode(email)
         }
     }
 
     private fun initPinCodeListener() {
         createPersonalAccountPinCodePinEditText.onTextCompleteListener = object : OnTextCompleteListener {
             override fun onTextComplete(code: String): Boolean {
-                createPersonalAccountWithEmailViewModel.activateEmail(email, code)
+                createPersonalAccountPinCodeViewModel.activateEmail(email, code)
                 return false
             }
         }
     }
 
     private fun observeActivateEmailData() {
-        with(createPersonalAccountWithEmailViewModel) {
+        with(createPersonalAccountPinCodeViewModel) {
             activateEmailSuccessLiveData.observe(viewLifecycleOwner) {
                 emailCredentialsViewModel.saveActivationCode(
                     createPersonalAccountPinCodePinEditText.text.toString()
@@ -77,7 +79,7 @@ class CreatePersonalAccountPinCodeFragment : Fragment(
     }
 
     private fun observeActivationCodeData() {
-        with(createPersonalAccountWithEmailViewModel) {
+        with(createPersonalAccountPinCodeViewModel) {
             sendActivationCodeSuccessLiveData.observe(viewLifecycleOwner) {
                 //TODO show correctly send activation code success messages
             }
@@ -103,7 +105,7 @@ class CreatePersonalAccountPinCodeFragment : Fragment(
     }
 
     private fun observeNetworkConnectionError() {
-        createPersonalAccountWithEmailViewModel.networkConnectionErrorLiveData.observe(viewLifecycleOwner) {
+        createPersonalAccountPinCodeViewModel.networkConnectionErrorLiveData.observe(viewLifecycleOwner) {
             showNetworkConnectionErrorDialog()
         }
     }
